// aceruapp/src/app/modules/problems/model/problema.ts
export class Problema {

    id?:              number;
    nombre?:          string;
    descripcion?:     string;
    entradas?:        string;
    salidas?:         string;
    ejemploEntradas?: string;
    ejemploSalidas?:  string;
    dificultad?:      string;
    estado?:          string;
    
}